[Helm Chart] Remove ServiceMonitor's namespaceSelector
Release.Namespace is specified in ServiceMonitor's namespaceSelector,
but the namespace actually installed is the kubelet-csr-approver.namespace.

Therefore, the ServiceMonitor will not work properly if it is installed
in a namespace other than the default namespace.

There is no need to specify a namespaceSelector because the
namespaceSelector defaults to the namespace to which the ServiceMonitor
belongs, which is also the namespace to which the Service belongs.

Signed-off-by: Mitsuru Kariya <mitsuru.kariya@nttdata.com>